back/requirements.txt
chromadb==0.4.13
chroma-hnswlib==0.7.3
python-docx==1.1.0
fastapi==0.104.1
nltk==3.8.1
pydantic==2.5.2
python-multipart==0.0.6
spacy
torch==2.0.1
torchaudio==2.0.2
uvicorn[standard]==0.24.0.post1
langchain==0.0.340
sentence-transformers
pysqlite3-binary
